+++ src/utils/binary_decoder.h
@@ -66,7 +66,7 @@ unsigned binary_decoder::next_1B() throw (binary_decoder_error) {
 
 unsigned binary_decoder::next_2B() throw (binary_decoder_error) {
   if (data + sizeof(uint16_t) > data_end) throw binary_decoder_error("No more data in binary_decoder");
-  unsigned result;
+  uint16_t result;
   memcpy(&result, data, sizeof(uint16_t));
   data += sizeof(uint16_t);
   return result;
@@ -74,7 +74,7 @@ unsigned binary_decoder::next_2B() throw (binary_decoder_error) {
 
 unsigned binary_decoder::next_4B() throw (binary_decoder_error) {
   if (data + sizeof(uint32_t) > data_end) throw binary_decoder_error("No more data in binary_decoder");
-  unsigned result;
+  uint32_t result;
   memcpy(&result, data, sizeof(uint32_t));
   data += sizeof(uint32_t);
   return result;
