+++ src/utils/pointer_decoder.h
@@ -45,14 +45,14 @@ unsigned pointer_decoder::next_1B() {
 }
 
 unsigned pointer_decoder::next_2B() {
-  unsigned result;
+  uint16_t result;
   memcpy(&result, data, sizeof(uint16_t));
   data += sizeof(uint16_t);
   return result;
 }
 
 unsigned pointer_decoder::next_4B() {
-  unsigned result;
+  uint32_t result;
   memcpy(&result, data, sizeof(uint32_t));
   data += sizeof(uint32_t);
   return result;
